requirements.txt
requests==2.31.0
pygame==2.5.2
python-dotenv==1.0.0
pyttsx3==2.90
TTS==0.22.0
torch==2.1.0
flask==3.0.0
flask-cors==4.0.0